Extract repro for #16739 into a separate file
